✨ feat: redis -> rdb로 조회수 옮기는 기능
- scheduler을 사용해서 3분마다 옮겨질 수 있도록 구현
- redis 내부 값들의 유효시간은 5분으로 설정
- redis의 incrementData 메서드를 통해 동기화 처리
  - synchronized 메서드를 통해 구현

+++ src/main/java/antigravity/product/domain/entity/Product.java
@@ -22,7 +22,7 @@ public class Product extends BaseTimeEntity {
     private String name;
     private BigDecimal price;
     private Integer quantity;
-    private Integer viewed;
+    private Long viewed;
     private LocalDateTime deletedAt;
     @Builder
     public Product(String sku, String name, BigDecimal price, Integer quantity, LocalDateTime deletedAt) {
@@ -32,7 +32,7 @@ public class Product extends BaseTimeEntity {
         this.quantity = quantity;
         this.deletedAt = deletedAt;
     }
-    void setViewed(Integer viewed) {
+    void setViewed(Long viewed) {
         this.viewed = viewed;
     }
